+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 # BetterBlizzPlates 2.1.2g
 ## Midnight
+### New
+- Add new Jaybeezy profile (www.twitch.tv/jaybeezy). Thank you for sharing!
 ### Tweak
 - Update Saul profile (www.twitch.tv/saul)
 - Update Snupy profile (www.twitch.tv/snupy)
+- Add a check for MiniAuras' nameplate settings and popup if overlap found and ask if you want to disable the similar BBP stuff.
+- Add a check for wrong CVars enabled with Totem Indicator and mention the limitations and fix it.
 ## Classics
 ### Bugfix
 - Fix level text perma showing 60 after the "midnight nameplates" update for classics.
+## All versions
+### Tweak
+- Add a hidden close button for the profiles sidebar in the top right corner of it that shows on mouseover.
 
 # BetterBlizzPlates 2.1.2f
 ## All Classics